# http/swagger.yml
openapi: "3.0.0"
info:
  title: Influx API Service
  version: 0.1.0
servers:
  - url: /v2
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: List tasks.
      description: Lists tasks, limit 100
      parameters:
        - in: query
          name: after
          schema:
            type: string
          description: returns tasks after specified ID
        - in: query
          name: user
          schema:
            type: string
          description: filter tasks to a specific user id
        - in: query
          name: organization
          schema:
            type: string
          description: filter tasks to a specific organization id
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tasks"
                  links:
                    $ref: "#/components/schemas/Links"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Tasks
      summary: Create a new task
      requestBody:
        description: task to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  '/tasks/{taskId}':
    get:
      tags:
        - Tasks
      summary: Retrieve an task
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: ID of task to get
      responses:
        '200':
          description: task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags:
        - Tasks
      summary: Update a task
      description: Update a task. This will cancel all queued runs.
      requestBody:
        description: task update to apply
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: ID of task to get
      responses:
        '200':
          description: task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: Deletes a task and all associated records
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: ID of task to delete
      responses:
        '204':
          description: task deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  '/tasks/{taskId}/runs':
    get:
      tags:
        - Tasks
      summary: Retrieve list of run records for a task
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: ID of task to get runs for
        - in: query
          name: after
          schema:
            type: string
          description: returns runs after specified ID
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: the number of runs to return
        - in: query
          name: afterTime
          schema:
            type: string
            format: date-time
          description: filter runs to those scheduled after this time, RFC3339
        - in: query
          name: beforeTime
          schema:
            type: string
            format: date-time
          description: filter runs to those scheduled before this time, RFC3339
      responses:
        '200':
          description: a list of task runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Run"
                  links:
                    $ref: "#/components/schemas/Links"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  '/tasks/{taskId}/runs/{runId}':
    get:
      tags:
        - Tasks
      summary: Retrieve a single run record for a task
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: task ID
        - in: path
          name: runId
          schema:
            type: string
          required: true
          description: run ID
      responses:
        '200':
          description: The run record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  '/tasks/{taskId}/runs/{runId}/retry':
    post:
      tags:
        - Tasks
      summary: Retry a task run
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: task ID
        - in: path
          name: runId
          schema:
            type: string
          required: true
          description: run ID
      responses:
        '200':
          description: The newly created retry run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  '/tasks/{taskId}/logs':
    get:
      tags:
        - Tasks
      summary: Retrieve all logs for a task
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: ID of task to get logs for
      responses:
        '200':
          description: all logs for a task
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Link"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  '/tasks/{taskId}/runs/{runId}/logs':
    get:
      tags:
        - Tasks
      summary: Retrieve all logs for a run
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: ID of task to get logs for.
        - in: path
          name: runId
          schema:
            type: string
          required: true
          description: ID of run to get logs for.
      responses:
        '200':
          description: all logs for a run
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Link"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Link:
      type: object
      readOnly: true
      description: URI of resource.
      properties:
        href:
          type: string
          format: url
      required: [href]
    Links:
      type: object
      properties:
        next:
          $ref: "#/components/schemas/Link"
        self:
          $ref: "#/components/schemas/Link"
        prev:
          $ref: "#/components/schemas/Link"
      required: [self]
    Organization:
      properties:
        id:
          readOnly: true
          type: string
        name:
          type: string
        owners:
          $ref: "#/components/schemas/Owners"
      required: [name]
    Organizations:
      type: array
      items:
        $ref: "#/components/schemas/Organization"
    Owners:
      properties:
        users:
          $ref: "#/components/schemas/Users"
        organizations:
          $ref: "#/components/schemas/Organizations"
    Run:
      properties:
        id:
          readOnly: true
          type: string
        taskId:
          readOnly: true
          type: string
        status:
          readOnly: true
          type: string
          enum: [
            "scheduled",
            "executing",
            "failed",
            "success"
          ]
        scheduledFor:
          description: Time used for run's "now" option, RFC3339.
          type: string
          format: date-time
        startedAt:
          readOnly: true
          description: Time run started executing, RFC3339Nano.
          type: string
          format: date-time
        finishedAt:
          readOnly: true
          description: Time run finished executing, RFC3339Nano.
          type: string
          format: date-time
        requestedAt:
          readOnly: true
          description: Time run was manually requested, RFC3339Nano.
          type: string
          format: date-time
        error:
          $ref: "#/components/schemas/Error"
        log:
          readOnly: true
          description: Link to the full logs for a run.
          type: string
    Task:
      properties:
        id:
          readOnly: true
          type: string
        organization:
          description: The ID of the organization that owns this Task.
          type: string
        name:
          description: A modifiable description of the task.
          type: string
        status:
          description: The current status of the task. When updated to 'disabled', cancels all queued jobs of this task.
          type: string
          enum: [
            "enabled",
            "disabled"
          ]
        owners:
          $ref: "#/components/schemas/Owners"
        flux:
          description: The Flux script to run for this task.
          type: string
        every:
          readOnly: true
          description: A simple task repetition schedule; parsed from Flux.
          type: string
        cron:
          readOnly: true
          description: A task repetition schedule in the form '* * * * * *'; parsed from Flux.
          type: string
        last:
          $ref: "#/components/schemas/Run"
      required: [name, organization, flux]
    Tasks:
      type: array
      items:
        $ref: "#/components/schemas/Task"
    User:
      properties:
        id:
          readOnly: true
          type: string
        name:
          type: string
      required: [name]
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User"
...
    Error:
      properties:
        code:
          readOnly: true
          type: integer
          format: int32
        message:
          readOnly: true
          type: string
      required: [code, message]
...